refactor: move mongodb to db
