+++ lib/db/account/writes.js
@@ -5,7 +5,7 @@ import {
   getPostsCollection,
   getStarsCollection,
   getUsersCollection,
-} from "../../mongodb";
+} from "../mongodb";
 import {
   getCommentsByUserId,
   getStarsByUserEmail,
