+++ lib/db/comments/reads.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getCommentsCollection } from "../../mongodb";
+import { getCommentsCollection } from "../mongodb";
 import { getPostById } from "../reads";
 
 /**
