+++ lib/db/comments/writes.js
@@ -3,7 +3,7 @@ import {
   getCommentsCollection,
   getInboxCollection,
   getPostsCollection,
-} from "../../mongodb";
+} from "../mongodb";
 import { getPostById, getUserByEmail, getUserById } from "../reads";
 import { v4 as uuid } from "uuid";
