+++ lib/db/inbox/reads.js
@@ -1,4 +1,4 @@
-import { getInboxCollection } from "../../mongodb";
+import { getInboxCollection } from "../mongodb";
 
 /**
  * Gets the inbox data of a user
