+++ lib/db/inbox/writes.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getInboxCollection } from "../../mongodb";
+import { getInboxCollection } from "../mongodb";
 
 /**
  * Marks some inbox messages as read.
