+++ lib/db/mongodb.js
@@ -1,5 +1,3 @@
-/** @module lib/mongodb */
-
 // This approach is taken from https://github.com/vercel/next.js/tree/canary/examples/with-mongodb
 import { MongoClient } from "mongodb";
 import { backOff } from "exponential-backoff";
