+++ lib/db/posts/reads.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getPostsCollection } from "../../mongodb";
+import { getPostsCollection } from "../mongodb";
 import { decryptPost } from "../../security/transformations";
 import { getUserByEmail } from "../reads";
