+++ lib/db/posts/writes.js
@@ -4,7 +4,7 @@ import {
   getCommentsCollection,
   getPostsCollection,
   getStarsCollection,
-} from "../../mongodb";
+} from "../mongodb";
 import { encryptPost } from "../../security/transformations";
 import { getPostById, getUserByEmail, hasAiCommentedOnPost } from "../reads";
 import { isPostOwner } from "../../validations";
