+++ lib/db/stars/reads.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getStarsCollection } from "../../mongodb";
+import { getStarsCollection } from "../mongodb";
 import { getPostById } from "../reads";
 
 export async function getStars(postId) {
