+++ lib/db/stars/writes.js
@@ -3,7 +3,7 @@ import {
   getInboxCollection,
   getPostsCollection,
   getStarsCollection,
-} from "../../mongodb";
+} from "../mongodb";
 import { getPostById, getUserByEmail, getUserById } from "../reads";
 import { getStar } from "./reads";
 import { v4 as uuid } from "uuid";
