+++ lib/db/users/reads.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getUsersCollection } from "../../mongodb";
+import { getUsersCollection } from "../mongodb";
 
 /**
  * Gets an user by its email
