+++ lib/db/users/writes.js
@@ -1,5 +1,5 @@
 import { ObjectID } from "bson";
-import { getUsersCollection } from "../../mongodb";
+import { getUsersCollection } from "../mongodb";
 
 /**
  * Updates a user's account with the provided data.
