+++ lib/db/writes.js
@@ -3,7 +3,7 @@ import { ObjectID } from "bson";
 import {
   getCompletionsCollection,
   getCosineSimilarityCollection,
-} from "../mongodb";
+} from "./mongodb";
 import { getUserByEmail } from "./reads";
 import OpenAI from "openai";
 import { hitChiron } from "../chiron";
